CS162 A03: Dynamic Memory + Class + Pointers
Using a File-in, File-out (FIFO) 
function(s), manipulate the date for the Fleet (class)
of Airplane (class) as they are nested Classes. Both 
classes need private and public member functions that
are dynamic for memory. Both classes require the 
"rule of 3": Destructor, Copy Constructor, & Overloaded
Assignment Operator. Beyond the dynamic memory 
addition to this version of the program, a growArray()
needs to be created that will be nested in Insert()
function. The program is menu-driven by the user. 
They can view planes, add a new plane, search for a
plane, etc.